backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.7
numpy==1.24.3
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
scikit-learn==1.3.2
phidata==2.4.11
